Refactor store of game query model #29
There's a new interface GameStore which is responsible for
saving and retrieving games. The interface GameFinder remains
because not every GameFinder can save a game.

The class GameProjection is now completely unaware of how a game
is stored. The in memory cache is gone too. The logic is now
in a dedicated class which also fixes the memory leak.