Squash event store followers #76
Each context registers a reusable command and
injects its own subscribers which can be selected.

The user of the command can set the pointer name.
The user of the command can select specifc subscribers.
The user of the command can set the batch size.
The user of the command can set the throttle time.