Improve integration with chat context (#74)
Implement non-blocking request-reply communication #73
The context that wants to initiate a chat can provide a
reply address (in this case its queue). The chat context
will then reply to this address so that the requester
doesn't have to process every generic ChatInitiated event.
Using the generic event would lead to unnecessary
database queries and/or message passing. However, the
possibility to listen to ChatInitiated still exists.

Unrelated changes
* Apply the same coding style (using match expression
instead of reflection) for all consumers. This way,
used private methods are not marked as unused.