Use uuid version 6 (#95)
Use uuid version 6 in all contexts #91
With version 6, the bits no longer need to be shifted to
create near ascending identities for the database. Drop
ramsey/uuid in favor of symfony/uid to keep the number
of vendors low. The doctrine type uuid_binary_ordered_time
is no longer needed.