Parallelize event store follower (#99)
Parallelize event store follower #98
Use pcntl_fork to distribute the load among the workers.
The order of the messages is maintained by using the
aggregate id as the shard id. The new '--worker' command
option defines the number of forks that are created.

Misc changes:
* The '-v' command option outputs the received events to stdout.
* If a non-existing subscriber is selected, the command exits.
* If no subscriber is selected, all available subscribers are shown.
* Remove ThrottlingEventStore and handle sleep directly
at the caller. Allows to decouple from synchronization.
* Replace StoredEventPublisher with StoredEventSubscriber
as this interface already helps to distribute the load
to the workers, see ForwardToChannelStoredEventSubscriber.
StoredEventPublisher itself is replaced by a composite.