Use web server as sidecar next to FPM (#131)
Use the script provided by the base image to run
Nginx and PHP-FPM in the same container. The Nginx
running in this container is not running as an edge router,
so the application can scale just as well as before,
just now at the HTTP level #115.

Application specific reasons for that change:
* Performance with Docker Swarm setups is improved
by avoiding another network hop to a potentially different node.
Unix sockets are used for IPC.
* Reduces TCP TIME_WAIT state on FPM containers. Connections
to FPM cannot be kept alive. TIME_WAIT state can rapidly grow on
high load or stress tests. The result is an unresponsive application.
* Allows to run a HTTP service in FPM mode to not have to
change the execution model of this PHP application.