Provide scheduler library
The scheduler can be used to interrupt the application
at a certain time to trigger certain handlers.
It's an abstraction around pcntl_alarm with the advantage
that multiple handlers can be registered.
For example, it can be used to send heartbeats to
applications like RabbitMQ or MySQL to maintain
connections in a long-running process.