Provide scheduler library (#132)
The scheduler can be used to interrupt the application
at a certain time to trigger certain handlers. It's an abstraction
around pcntl_alarm with the advantage that multiple handlers
can be registered. For example, heartbeats can be sent to
applications such as RabbitMQ or MySQL to maintain
connections in a long-running process without having to
switch the execution model (e.g. fibers) of those processes.