Remove hasUncommittedStoredEventId
This was used to detect gaps in the sequence of
stored event ids. It was already misplaced in the old
design. The method (and the decorator) is replaced
by StoredEventFilters::untilGapIsFound() which accepts
various strategies on how to handle gaps.

Furthermore, UnrecoverableException is removed. The intention
was good, but this explicit exception wasn't needed in the
years.