Segregate EventStore interface (#140)
To feed stored event subscribers, a second interface
PollableEventStore is introduced. Not every implementation
of EventStore needs to be pollable, as others may tail the
transaction log of the underlying database #79.

Additional improvements:
* hasUncommittedStoredEventId is removed from the
interface. It only existed to potentially wait for gaps to be
filled. This was provided by a decorator, which in retrospect
was not the right abstraction. This functionality is now explicitly
offered via StoredEventFilters::untilGapIsFound() that accepts
a GapDetection strategy.
* UnrecoverableException is removed. The intention
was good, but this explicit exception wasn't needed in the
last years.
* NormalizerException is removed from public api. That's
an implementation detail. It was also never thrown in the
current implementations. It got caught and re-thrown as
an EventStoreException.