Use Symfony's security system
Upon arrival, ArrivialAuthenticator creates a new user in the
identity context and authenticates them in the session.

The current implementation covers what AssignUserIdOnKernelRequest
did before. It adds some complexity, but opens the way for #130
to secure the /metrics endpoint with basic auth, and #6 to eventually
authenticate the user in any way.