Update dependencies
The signature of Doctrine\DBAL\Driver::connect() got more strict.
The analyzer complained that non-existing keys are used, in this
case for heartbeat. Ideally the key driverOptions should be used
to configure the driver. But since these options are used
one-to-one for the decorated driver, it fails because it doesn't
know the options. We'd have to clean up, which is also done in
Parameters::removeDriverOptions(). Unfortunately, the middleware
isn't called in all cases and the options slip through anyway,
for example at doctrine:database:create. That's why the
heartbeat option is still accessed from the root level, but
encapsulated in the Parameters class.