Cleanup environment variables
* Combine environment variables
Some environment variables were defined in docker-compose.yml
as they were used within the `entrypoint`. The environment
file was only loaded within Symfony. That changes with this
commit as all variables are defined in that file. This
mostly affects the dev and pipeline environment. With that
change, also the Dotenv component from Symfony is removed.
In prod, everything will be declared anyways, and in dev
the environment file doesn't change that frequently.
* Replace APPLICATION prefix with APP
It's just shorter. Additionally all environment variables
are now prefixed with APP.
* Move environment file to root folder
This is a followed standard by most projects.
* Do cache warmup in entrypoint
This was done during image build time. However, environment
variables aren't available at build time anymore.