Enable acceptance tests
* Besides the repository, the event store access must also be
sharding aware, since this is read when the query model is not
yet populated. For that matter, to query the events from the
write model, the repository is used, which encapsulates the
event store. Additionally, the event store is used directly
for appending events.
* Introduce ShardChooser
ShardChooser encapsulates the logic of selecting a shard.
* Remove query model's GameNotFoundException
This is an aside refactoring, not related. This indirection
isn't needed.