Secure metrics via ip addresses
The environment variable APP_METRICS_ALLOWED_IPS
can take a comma-separated list of ip addresses.
Additionally, APP_TRUSTED_PROXIES can be configured.
This is needed because when run, e.g., in a docker
environment, the allowed ips for metrics can cover
the whole ip address range of docker. If the load balancer
runs within the same range, it would always be allowed
to access the metrics endpoint. With trusted proxies,
it would forward the client's ip address, which at the end
would be rejected.