Add ProxySQL (#167)
Adds ProxySQL to the stack, which acts as a connection pool
and provides the ability to route queries to different physical
MySQL machines, e.g. for schema-based sharding #119.

Applications communicate with ProxySQL via unix domain sockets,
which means that it will be deployed as a sidecar.
There are no query rules defined for the environments yet.
ProxySQL will pay of during load tests #118.

ProxySQL is optionally configurable via the PROXYSQL_CONFIG
environment variable, which is required for the Play with Docker
environment because the stack file must be self-contained.