Initiate chats exactly once
* Whilst issuing a command to initiate a chat, downstream contexts
need to attach an idempotency key that must be globally unique.
The chat context will make sure that the same idempotency keys
produce the same results. Although idempotent, the chat context is
configured to clean up idempotency keys after 1 day. Within that
timeframe, duplicates are handled gracefully. That's the SLA for
up- and downstreams.
* Create abstraction for an idempotent storage and provide a predis
implementation.
* Create new environment variable for predis for the chat context.