Fix segfault in dev environment
The process removal in Wait::any() caused a segfault
in the dev environment when a child process got
terminated. The reason is not discovered yet. Adding
all parameters, assigning the return value to a
variable before passing it to Processes:remove() or
extracting pcntl_wait to a function fixes that issue.