Add hooks via PSR-14
Implementations are supposed to dispatch the defined events
within the Gaming\Common\MessageBroker\Event namespace.
Currently it's only used for logging, but it could also
be used for various other enhancements, such as resetting
the container after a message has been handled.

For logging specifically, this design is an improvement
compared to the decorator that was used before as each
implementation (in this case AmqpLib) can give more context,
such as the queue, the exchange and the routing key.