Make AmqpPublisher reliable by default
Remove the ReliablePublishing abstraction and automatically enable
publisher confirms. They're also not that slow, so this feature
makes a good default. It may come back later as removing it
violates OCP, but for now it simplifies the design. Furthermore,
the implementation is now really reliable which is achieved by
dispatching MessageReturned. Lost messages used to go unnoticed.