Add dynamic page loads
All links on the page are intercepted and checked to
see if they're a valid internal link. If so, the page
is loaded dynamically via ajax, then all undefined
custom elements are loaded to avoid flickering, and then
the page is placed in the DOM. The same works for popstate
events and programmatic triggers, such as joining a game.