Use Redis pipelining for projections (#226)
Pipelining reduces the round-trip time for each command,
resulting in higher throughput per worker. In addition, this
reduces the number of workers for the load test (#170) from
8 to 3, which leads to better resource efficiency #225.

Covers only the game projection. The added code complexity
isn't worthwhile for the other projections as they don't generate
such high network traffic. The default pipeline size is 32, as this
achieves the best results.