Call identity context in UserProvider (#237)
This saves the most important user data in the session.

As the UserProvider is based on an external API (not really, but the
design is aiming for that), the user is only refreshed at a configured
interval (default is 60s). This can cause the data in the session to
be stale, e.g. the username. If critical data is changed by the user
themselves, the security's user must be refreshed manually.

With a refresh interval of 60s, no performance is lost when testing
with #170 (same test scenarios). 0s shows a loss of ~3k req/s.