Toggle resign and abort buttons (#276)
The button states are calculated on the client. The business rule is
pretty fix, and the state only toggles once in a game. Duplicating
the business logic on the client, rather than the server tells the
client what to do is the more sane option in this scenario. Overall,
this produces less code, less data with the downside of being a bit
less DRY for the business logic.

To get the server scenario right, PlayerMoved would need to expose
abortable and resignable.