Make whole column interactive (#279)
This adds a new element for the tokens so the pseudo-element on the
field can be used to expand the click area #13. Another benefit is,
that the token can now be used as a standalone element. When the
token is part of a game field, the border width is calculated based
on the number of columns. Tokens can be used as a loading bar within
a .gp-cf-token-loading element.

Additionally, prefix CSS classes with gp-cf- (the cf is new) to avoid
conflicts with future games.